backend/requirements.txt
Django==2.2.19
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
gunicorn==20.0.4
Pillow==9.0.0
psycopg2-binary==2.8.6
python-dotenv==0.19.0